requirements.txt
numpy
scipy
matplotlib
pyvisa
pytest
pandas
openpyxl
jupyter
scikit-learn
pyserial
requests
flask
PyPNA
pyyaml